Introduced a version of mps_arena_formatted_objects_walk.
This version takes into account the fact that the stepper function might alter
the memory passed to it, and makes sure to simulate a barrier hit to update the
internal segment summaries. Otherwise, this could cause the MPS to fail to
update some references when moving an object in the future.

As such, the new function is called mps_arena_formatted_objects_walk_rw.